requirements.txt
python-telegram-bot
python-dotenv
dnspython
python-whois
requests
flask
phonenumbers
openai